Revert "Add disk cache store (#949)"
This reverts commit f86a376b0d088c8ff83ef8180b95d492527902e5.
