+++ src/core/cache.js
@@ -1,9 +1,8 @@
 import { setMetaContent } from "../util"
-import { SnapshotCache } from "./drive/snapshot_cache"
 
 export class Cache {
   clear() {
-    this.store.clear()
+    this.session.clearCache()
   }
 
   resetCacheControl() {
@@ -16,18 +15,6 @@ export class Cache {
 
   exemptPageFromPreview() {
     this.#setCacheControl("no-preview")
-  }
-
-  set store(store) {
-    if (typeof store === "string") {
-      SnapshotCache.setStore(store)
-    } else {
-      SnapshotCache.currentStore = store
-    }
-  }
-
-  get store() {
-    return SnapshotCache.currentStore
   }
 
   #setCacheControl(value) {
