+++ src/core/drive/page_snapshot.js
@@ -45,10 +45,6 @@ export class PageSnapshot extends Snapshot {
     return this.documentElement.getAttribute("lang")
   }
 
-  get html() {
-    return `${this.headElement.outerHTML}\n\n${this.element.outerHTML}`
-  }
-
   get headElement() {
     return this.headSnapshot.element
   }
