+++ src/core/drive/visit.js
@@ -215,8 +215,8 @@ export class Visit {
     }
   }
 
-  async getCachedSnapshot() {
-    const snapshot = (await this.view.getCachedSnapshotForLocation(this.location)) || this.getPreloadedSnapshot()
+  getCachedSnapshot() {
+    const snapshot = this.view.getCachedSnapshotForLocation(this.location) || this.getPreloadedSnapshot()
 
     if (snapshot && (!getAnchor(this.location) || snapshot.hasAnchor(getAnchor(this.location)))) {
       if (this.action == "restore" || snapshot.isPreviewable) {
@@ -235,8 +235,8 @@ export class Visit {
     return this.getCachedSnapshot() != null
   }
 
-  async loadCachedSnapshot() {
-    const snapshot = await this.getCachedSnapshot()
+  loadCachedSnapshot() {
+    const snapshot = this.getCachedSnapshot()
     if (snapshot) {
       const isPreview = this.shouldIssueRequest()
       this.render(async () => {
