--- src/tests/fixtures/disk_cache.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html id="html" data-skip-event-details="turbo:submit-start turbo:submit-end">
-  <head>
-    <meta charset="utf-8">
-    <meta name="csp-nonce" content="123">
-    <link rel="icon" href="data:;base64,iVBORw0KGgo=">
-    <title>Turbo</title>
-    <script src="/dist/turbo.es2017-umd.js" data-turbo-track="reload"></script>
-    <script src="/src/tests/fixtures/test.js"></script>
-    <script>
-      Turbo.cache.store = "disk"
-
-      document.addEventListener("turbo:load", async () => {
-        await new Promise(resolve => setTimeout(resolve, 100))
-
-        const cachesList = document.getElementById("caches")
-
-        const cache = await caches.open("turbo-v1")
-        const keys = await cache.keys()
-        cachesList.innerHTML = keys.map(key => `<li>${key.url}</li>`).join("")
-
-        const clearCacheButton = document.getElementById("clear-cache")
-        clearCacheButton.addEventListener("click", async (event) => {
-          await Turbo.cache.clear()
-          cachesList.innerHTML = ""
-        })
-      })
-    </script>
-  </head>
-  <body>
-    <h1>Cached pages:</h1>
-    <ul id="caches"></ul>
-
-    <h3>Links:</h3>
-    <ul>
-      <li>
-        <a id="first-link" href="./disk_cache.html">First HTTP cached page</a>
-      </li>
-      <li>
-        <a id="second-link" href="./disk_cache.html?page=2">Second HTTP cached page</a>
-      </li>
-      <li>
-        <a id="third-link" href="./disk_cache.html?page=3">Third HTTP cached page</a>
-      </li>
-    </ul>
-
-    <button id="clear-cache">Clear cache</button>
-  </body>
-</html>
